[Repositories] Add more precise types to repository generator (#2391)
* Make utils/scripts/generators/repository-generator.pl aware of more
datatypes

This adds support for unsigned and more integer types. It also avoids
using parsing functions that require casting (still needed in some
cases)

Having the data types in the Repository structs better map to the types
in the database will allow us to avoid casting when we pull data out of
them. And as a benefit, assume something is wrong if we do :)

Hopefully clean up some warnings due to casting too.

+++ common/repositories/base/base_alternate_currency_repository.h
@@ -19,8 +19,8 @@
 class BaseAlternateCurrencyRepository {
 public:
 	struct AlternateCurrency {
-		int id;
-		int item_id;
+		int32_t id;
+		int32_t item_id;
 	};
 
 	static std::string PrimaryKey()
@@ -118,8 +118,8 @@ public:
 		if (results.RowCount() == 1) {
 			AlternateCurrency e{};
 
-			e.id      = atoi(row[0]);
-			e.item_id = atoi(row[1]);
+			e.id      = static_cast<int32_t>(atoi(row[0]));
+			e.item_id = static_cast<int32_t>(atoi(row[1]));
 
 			return e;
 		}
@@ -242,8 +242,8 @@ public:
 		for (auto row = results.begin(); row != results.end(); ++row) {
 			AlternateCurrency e{};
 
-			e.id      = atoi(row[0]);
-			e.item_id = atoi(row[1]);
+			e.id      = static_cast<int32_t>(atoi(row[0]));
+			e.item_id = static_cast<int32_t>(atoi(row[1]));
 
 			all_entries.push_back(e);
 		}
@@ -268,8 +268,8 @@ public:
 		for (auto row = results.begin(); row != results.end(); ++row) {
 			AlternateCurrency e{};
 
-			e.id      = atoi(row[0]);
-			e.item_id = atoi(row[1]);
+			e.id      = static_cast<int32_t>(atoi(row[0]));
+			e.item_id = static_cast<int32_t>(atoi(row[1]));
 
 			all_entries.push_back(e);
 		}
